Various misc additions
Added frame.html and a style page for it. It's not very consequential,
it's just a little page to practice some html/styling stuff on,
I wanted to learn how to use iframes with it.

I edited some text and comments throughout index.html, data.js,
hockey_data.ejs. I removed some unnecessary code from server.js that I
forgot to delete in the copy-over from the original project.

I made some small additions to the readme

Most notably, I roster_search.html, which is a form very similar to
nhl_query.html, but it takes a number input to return the current
roster of players corresponding to a team ID in the NHLs API system.