Primarily progress in the /teams POST interaction
I built the desired interaction for /teams POST, receiving, formatting
and sending the file with the appropriate data to the user. Added
a way to display all the data categories to the User below it. In doing
so, I've added an "additional info" section to the body of the .ejs file

I created a static html page to contain the interactions to return
lists of teams/franchises. It's not properly styled, a WIP at present.

And I expanded on and updated the readme file