Introducing frontend javascript into this project
Hit a snag with functionality that I wanted to implement from the
backend, so now I'm tinkering with a potential frontend solution.
Created a new ejs file (might not even be necessary but I wanted
a blank canvass to mess around with) named player_search_template.ejs
(might simplify the name if I keep it). I created a scripts folder and
with a script file that is linked to this. Also have a corresponding
stylesheet. Created an extension to display it with in the data route